Add files via upload
Updated my code in this file to print a dictionary of items the user hypothetically bought with their corresponding prices and the total price like a receipt, instead of only the total price. Also, I added user feedback in the code of this file in case the user enters unusable input.